Move Slur::avoidCollisions to SlurTieLayout
