+++ src/engraving/dom/slur.h
@@ -57,7 +57,6 @@ public:
 
     Slur* slur() const { return toSlur(spanner()); }
     Shape getSegmentShape(Segment* seg, ChordRest* startCR, ChordRest* endCR);
-    void avoidCollisions(PointF& pp1, PointF& p2, PointF& p3, PointF& p4, mu::draw::Transform& toSystemCoordinates, double& slurAngle);
 
     const PointF& endPointOff1() const { return m_endPointOff1; }
     const PointF& endPointOff2() const { return m_endPointOff2; }
